// packages/openapi-bench/generators/ng-openapi-gen/client/fn/params/query-styles.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { Echo } from '../../models/echo';
import { PetFilter } from '../../models/pet-filter';

export interface QueryStyles$Params {
  required: string;
  tagsExplode?: Array<string>;
  tagsCsv?: Array<string>;
  tagsPipe?: Array<string>;
  tagsSpace?: Array<number>;
  filter?: PetFilter;
  point?: {
'x'?: number;
'y'?: number;
};
  since?: string;
  day?: string;
  flag?: boolean;
  sort?: 'asc' | 'desc';
}

export function queryStyles(http: HttpClient, rootUrl: string, params: QueryStyles$Params, context?: HttpContext): Observable<StrictHttpResponse<Echo>> {
  const rb = new RequestBuilder(rootUrl, queryStyles.PATH, 'get');
  if (params) {
    rb.query('required', params.required, {});
    rb.query('tagsExplode', params.tagsExplode, {"style":"form","explode":true});
    rb.query('tagsCsv', params.tagsCsv, {"style":"form","explode":false});
    rb.query('tagsPipe', params.tagsPipe, {"style":"pipeDelimited","explode":false});
    rb.query('tagsSpace', params.tagsSpace, {"style":"spaceDelimited","explode":false});
    rb.query('filter', params.filter, {"style":"deepObject","explode":true});
    rb.query('point', params.point, {"style":"form","explode":true});
    rb.query('since', params.since, {});
    rb.query('day', params.day, {});
    rb.query('flag', params.flag, {});
    rb.query('sort', params.sort, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<Echo>;
    })
  );
}

queryStyles.PATH = '/params/query';