// packages/openapi-bench/generators/ng-openapi-gen/client/fn/bodies/upload-files.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { UploadMeta } from '../../models/upload-meta';
import { UploadResult } from '../../models/upload-result';

export interface UploadFiles$Params {
      body: {
'title': string;
'file': Blob;
'attachments'?: Array<Blob>;
'meta'?: UploadMeta;
}
}

export function uploadFiles(http: HttpClient, rootUrl: string, params: UploadFiles$Params, context?: HttpContext): Observable<StrictHttpResponse<UploadResult>> {
  const rb = new RequestBuilder(rootUrl, uploadFiles.PATH, 'post');
  if (params) {
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<UploadResult>;
    })
  );
}

uploadFiles.PATH = '/bodies/multipart';